working with  Drug

+++ service/src/main/java/com/bottomUp/domain/DoctorDrugDoseSettingData.java
@@ -5,25 +5,25 @@ import com.bottomUp.domain.common.BaseData;
 
 public class DoctorDrugDoseSettingData  extends BaseData {
 
-  private long drugDoseSettingID;
-  private long drugSettingID;
+  private Long drugDoseSettingID;
+  private Long drugSettingID;
   private String dose;
-  private long numOfDay;
-  private long durationType;
+  private Integer numOfDay;
+  private Integer durationType;
 
-  public long getDrugDoseSettingID() {
+  public Long getDrugDoseSettingID() {
     return drugDoseSettingID;
   }
 
-  public void setDrugDoseSettingID(long drugDoseSettingID) {
+  public void setDrugDoseSettingID(Long drugDoseSettingID) {
     this.drugDoseSettingID = drugDoseSettingID;
   }
 
-  public long getDrugSettingID() {
+  public Long getDrugSettingID() {
     return drugSettingID;
   }
 
-  public void setDrugSettingID(long drugSettingID) {
+  public void setDrugSettingID(Long drugSettingID) {
     this.drugSettingID = drugSettingID;
   }
 
@@ -35,19 +35,19 @@ public class DoctorDrugDoseSettingData  extends BaseData {
     this.dose = dose;
   }
 
-  public long getNumOfDay() {
+  public Integer getNumOfDay() {
     return numOfDay;
   }
 
-  public void setNumOfDay(long numOfDay) {
+  public void setNumOfDay(Integer numOfDay) {
     this.numOfDay = numOfDay;
   }
 
-  public long getDurationType() {
+  public Integer getDurationType() {
     return durationType;
   }
 
-  public void setDurationType(long durationType) {
+  public void setDurationType(Integer durationType) {
     this.durationType = durationType;
   }
 }
